Move out the guts of cmds/bulk_mark to bulk_operations
